ClipboardSync.Python/requirements.txt
Pillow>=10.0.0
pywin32>=306
pystray>=0.19.5
